+++ apps/web/app/[workspaceSlug]/(dashboard)/chat/page.tsx
@@ -1,0 +1,1 @@
+export { ChatPage as default } from "@multica/views/chat";
