+++ packages/core/paths/consistency.test.ts
@@ -22,6 +22,7 @@ describe("paths.workspace() shape", () => {
         "autopilots",
         "agents",
         "inbox",
+        "chat",
         "myIssues",
         "runtimes",
         "skills",
@@ -40,6 +41,7 @@ describe("paths.workspace() shape", () => {
       ["autopilots", "autopilots"],
       ["agents", "agents"],
       ["inbox", "inbox"],
+      ["chat", "chat"],
       ["myIssues", "my-issues"],
       ["runtimes", "runtimes"],
       ["skills", "skills"],
