+++ packages/views/chat/index.ts
@@ -1,2 +1,1 @@
-export { ChatFab } from "./components/chat-fab";
-export { ChatWindow } from "./components/chat-window";
+export { ChatPage } from "./components/chat-page";
